backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
google-generativeai==0.7.0
google-cloud-discoveryengine==0.13.2
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
